Added Scale Buttons to menu; Added Yell property
Scale Buttons allow the entire screen to be bigger (for those of you with larger monitors). The yell property allows games to ask the framework to change in some way. This is checked by the framework during the update method and only considers certain games (right now, only the menu game).