Added Game Buttons and PreReq
Game buttons allow you to play a single game an infinite amount of times. PreReq is a function that gives games data before gload, so that games can read the game_dir (name) and bindings of another game. This was required to get the game names to print in the main menu.